requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
pytz>=2024.1
boto3>=1.34.0
